update query schema to support 3 state system
Previously, if a result was being truncated for having > 500 rows, I would append a notice to the resulting HTML table and hash all of it together.  It is better to only hash the resulting table, really.  (Ideally, I'd hash the raw result set object or something.)

So, this commit gets rid of that, and adds two new attributes to the Query model, truncated_results (bool) and result_size (int).