first efforts at actual db security!
The details of this commit took far longer to tease out than they could have.

I mistakenly believed for several days that I would need to have a multi-database schema.  In truth, it seems I only need a single database, but with multiple connections to it, each with varied privilege levels.

So, this currently expects there to be:
	1 database "ScriptKitty"
	2 database users "root" "thesis_learner"

Database migrations and schema should thus remain untouched and working.

Now to actually test the security of this system!