check for JOIN statements when parsing table names
I didn't think anyone would use JOINs in Script Kitty, but Jason quickly proved that assumption wrong!  My former parse_out_tables() method didn't check for JOINs which caused other problems down the line when only one table was found in the FROM clause.

This commit checks for the possibility of multiple JOINs and multiple tables in the FROM clause and ultimately returns an array of all tables detected in the learner's query.