update TODO
The admin panel seems to be done for now.

The educator panel is established, and can be improved upon after I talk with Simon about learner metrics.

Sharing the Certificate of Completion™ to social media platforms should probably be the next thing I focus on as it directly impacts learners.

Additional exercises are optional, really, for now.